Extract `convertMarkupToJSON` into own module
The `Call` wrapper probably should focus on the responsibility of CLI
dispatch, not the deeper MarkupToJson logic.

There's a bug here though...
